Separated StopInfo.cs
